+++ cognite/client/_api/datapoints_subscriptions.py
@@ -50,13 +50,12 @@ class DatapointsSubscriptionAPI(APIClient):
                 >>> created = client.time_series.subscriptions.create(sub)
 
             Create a filter defined subscription for all numeric time series:
-
                 >>> from cognite.client import CogniteClient
                 >>> from cognite.client.data_classes import DataPointSubscriptionWrite
                 >>> from cognite.client.data_classes import filters as flt
-                >>> from cognite.client.data_classes.datapoints_subscriptions import DatapointSubscriptionFilterProperties
+                >>> from cognite.client.data_classes.datapoints_subscriptions import DatapointSubscriptionProperty
                 >>> client = CogniteClient()
-                >>> prop = DatapointSubscriptionFilterProperties.is_string
+                >>> prop = DatapointSubscriptionProperty.is_string
                 >>> numeric_timeseries = flt.Equals(prop, False)
                 >>> sub = DataPointSubscriptionWrite(
                 ...     "mySubscription",